Split UseHooks trait in several files
Adding one Trait by hook to improve readability, and preparing to scalability to be able to use more hooks in the future.
